requirements.txt
discord.py==2.6.4